refactor(actors): split llms.py into llms, openai, and genai
Move the concrete provider backends out of actors/llms.py into their own
modules: OpenAI (with its _parse_think_tags helper) to actors/openai.py and
GoogleGenAI to actors/genai.py. actors/llms.py keeps the abstract LLMChat,
LLMResponse, ReasoningLevel and the shared _extract_extra_usage_metadata
helper, and re-exports OpenAI/GoogleGenAI at the bottom so existing imports
(e.g. `from kaggle_benchmarks.actors.llms import OpenAI, GoogleGenAI`) keep
working. Absolute imports mean `import openai` inside actors/openai.py still
resolves to the third-party package (same pattern as serializers/openai.py).

test_openai_client imports _parse_think_tags from its new home
(actors.openai); DEVELOPMENT.md / AGENTS.md note the new layout.
